refactor: arrange import paths for PeerState
