+++ src/services/cluster/actors/mod.rs
@@ -1,3 +1,2 @@
 pub(super) mod actor;
-pub(crate) use crate::services::cluster::replication::replication::PeerState;
-pub(crate) mod listening_actor;
+pub(super) mod listening_actor;
