+++ src/services/cluster/command/cluster_command.rs
@@ -1,8 +1,7 @@
-use crate::services::cluster::actors::PeerState;
 use crate::services::cluster::peer::address::PeerAddrs;
 use crate::services::cluster::peer::identifier::PeerIdentifier;
 use crate::services::cluster::peer::kind::PeerKind;
-use crate::services::cluster::replication::replication::Replication;
+use crate::services::cluster::replication::replication::{PeerState, Replication};
 use crate::services::query_io::QueryIO;
 use tokio::net::TcpStream;
